ПР feature 30: отложенные сообщения (rev.1)
